requirements.txt
annotated-types==0.7.0
argcomplete==1.10.3
asgiref==3.9.1
beautifulsoup4==4.8.2
cachetools==5.5.2
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
colorama==0.4.6
compressed-rtf==1.0.7
Django==5.2.4
djangorestframework==3.16.0
docx2txt==0.9
ebcdic==1.1.1
extract-msg==0.28.7
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
IMAPClient==2.1.0
lxml==6.0.0
numpy==2.2.6
olefile==0.47
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==0.6.23
requests==2.32.4
rsa==4.9.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
sqlparse==0.5.3
textract==1.6.5
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
xlrd==1.2.0
xlsxwriter==3.2.5
